Support Prometheus metrics directly (#29)
Prometheus requires pre-declaring all metrics and their labels, and for a given metric name only one set of labels is allowed, otherwise panic occurs.

The approach here is to cache metric vectors registered with Prometheus using metric's name and a sorted list of labels as the cache key. This allows creating Jaeger metrics objects multiple times even with the same name and a set of tags.