metrics: break out local backend into metrics/testutils/metricstest
The local backend is only supposed to be used by tests, but because it
was part of the metrics package, it got pulled into all applications
using jaeger-client-go, increasing code size that must be vendored.

In addition, it also pulled in an unmaintained
component (https://github.com/codahale/hdrhistogram).

Now the local backend and the metrics utility code are in a new
metrics/testutils/metricstest package that is not pulled in via the
client API. Beware that this an API break because "metrics" no longer
has the same functions.

To minimize the API change, GetKey() is kept in "metrics" because it
doesn't depend on the local backend.

To break a dependency cycle (metrics -> local -> metrics),
metrics_test.go has to be in a metrics_test package and do black-box
testing. This in turn made it necessary to export initMetrics as part
of the metrics API. It was renamed to InitOrError and documented
accordingly to emphasize that it has the same functionality and only
differs in the return value.

Unfortunately the logic is now backwards compared to other packages
(regexp.Compile returns an error, regexp.MustCompile panics), but
preserving API compatibility with older releases is more important.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>